fix lambda and web to work with new api

+++ web/src/app/services/test.service.ts
@@ -14,6 +14,7 @@ import { saveAs } from 'file-saver';
 import { CustomHttpParamEncoder } from './encoder';
 import { ValueQuestion } from '../model/value-question';
 import { SolvableValueQuestion } from '../model/solvable-value-question';
+import {AuthenticationRecruiterService} from './authentication-recruiter.service';
 
 const httpOptions = {
   headers: new HttpHeaders({
@@ -27,19 +28,20 @@ const httpOptions = {
 
 export class TestService {
   constructor(
-    private httpClient: HttpClient
+    private httpClient: HttpClient,
+    private authService: AuthenticationRecruiterService
   ) { }
 
-  private testUrl: string = Globals.apiBaseUrl + '/recruiters/tests';
+  private testUrl: string = Globals.apiBaseUrl + '/recruiters';
 
   public createTest(inputTestTitle, author, language, openQuestions, closeQuestions, valueQuestions) {
-    var test = new Test('', inputTestTitle, author, language, openQuestions, closeQuestions, valueQuestions);
-    return this.httpClient.post<Test>(this.testUrl, test, httpOptions);
+    var test = new Test(this.authService.getUserId(), null, inputTestTitle, language, openQuestions, closeQuestions, valueQuestions);
+    return this.httpClient.post<Test>(`${this.testUrl}/${this.authService.getUserId()}/tests`, test, httpOptions);
   }
 
-  public updateTest(id, inputTestTitle, author, language, openQuestions, closeQuestions, valueQuestions) {
-    var test = new Test(id, inputTestTitle, author, language, openQuestions, closeQuestions, valueQuestions);
-    return this.httpClient.patch(this.testUrl, test, httpOptions);
+  public updateTest(testId, inputTestTitle, author, language, openQuestions, closeQuestions, valueQuestions) {
+    var test = new Test(this.authService.getUserId(), testId, inputTestTitle, language, openQuestions, closeQuestions, valueQuestions);
+    return this.httpClient.put(`${this.testUrl}/${this.authService.getUserId()}/tests/${testId}`, test, httpOptions);
   }
 
   public async translateTest(test: Test, language: string) {
@@ -50,7 +52,7 @@ export class TestService {
     var yandexKey = 'trnsl.1.1.20200108T191910Z.fe657624420b3a8c.9b1c3b15e8688d96a425d4596dfc2c6321f04ee2';
     var translateUrl = 'https://translate.yandex.net/api/v1.5/tr.json/translate?key=';
     var lang = ''
-    var result = new Test('', '', '', '', [], [], []);
+    var result = new Test('', null, '', '', [], [], []);
     if (language == 'pl') {
       lang = '&lang=en-pl';
     } else {
@@ -96,7 +98,7 @@ export class TestService {
     }
 
     console.log(result.openQuestions);
-    return this.httpClient.post<Test>(this.testUrl, result, httpOptions);
+    return this.httpClient.post<Test>(`${this.testUrl}/${this.authService.getUserId()}/tests/`, result, httpOptions);
   }
 
   public downloadTest(test: Test): void {
@@ -146,7 +148,7 @@ export class TestService {
     console.log('plik csv' + csv);
 
     let file = new Blob([csv], { type: 'text/csv;charset=utf-8' });
-    saveAs(file, test.title + '-' + test.id + '.csv');
+    saveAs(file, test.title + '-' + test.testId + '.csv');
   }
 
   public importTest(file: string) {
@@ -174,24 +176,24 @@ export class TestService {
     });
     language = splitFile[0].split(';')[2];
 
-    var test = new Test('', '', '', language, openQuestions, closeQuestions, valueQuestions);
+    var test = new Test('', null, '', language, openQuestions, closeQuestions, valueQuestions);
     console.log(test);
     return test;
-    // return this.httpClient.post<Test>(this.testUrl, test, httpOptions);
+    // return this.httpClient.post<Test>(`${this.testUrl}/${this.authService.getUserId()}/tests/`, test, httpOptions);
   }
 
-  public getTest(testID: string) {
+  public getTest(testID: number) {
     var httpOptions = {
       headers: new HttpHeaders({
         'Content-Type': 'application/json'
       })
     };
-    return this.httpClient.get(this.testUrl + `/${testID}`, httpOptions);
+    return this.httpClient.get(`${this.testUrl}/${this.authService.getUserId()}/tests/${testID}`, httpOptions);
   }
 
   public getAllTests(title: string = ''): Observable<Test[]> {
     const params = new HttpParams({ encoder: new CustomHttpParamEncoder() }).set('title', title);
-    return this.httpClient.get<Test[]>(this.testUrl, { params });
+    return this.httpClient.get<Test[]>(`${this.testUrl}/${this.authService.getUserId()}/tests/`, { params });
   }
 
   public getAllUserTests(username: string) {
@@ -240,15 +242,14 @@ export class TestService {
     return this.getAllUserTests(username).pipe(map(tests => tests.find(test => test.timestamp === timestamp)));
   }
 
-  public deleteTest(test: Test) {
+  public deleteTest(testId: number) {
     var httpOptions = {
       headers: new HttpHeaders({
         'Content-Type': 'application/json'
-      }),
-      body: test.id
+      })
     };
 
-    return this.httpClient.delete(this.testUrl, httpOptions);
+    return this.httpClient.delete(`${this.testUrl}/${this.authService.getUserId()}/tests/${testId}`, httpOptions);
   }
 
   public deleteTestInstance(applicantId: string, timestamp: string) {
@@ -257,7 +258,7 @@ export class TestService {
 
   public sendSolvedTest(test: TestInstance) {
     return new Observable(observer => {
-      this.httpClient.put<TestInstance>(Globals.apiBaseUrl + '/applicants/' + test.applicantID + '/tests', test, httpOptions)
+      this.httpClient.put<TestInstance>(Globals.apiBaseUrl + '/applicants/' + test.applicantId + '/tests', test, httpOptions)
         .subscribe({
           error: err => {
             console.log(err);
@@ -269,5 +270,9 @@ export class TestService {
           }
         });
     });
+  }
+
+  assignApplicantToTest(_testId: string, applicantId: string, confirm: boolean) {
+    return this.httpClient.post<string>(`${Globals.apiBaseUrl}/applicants/${applicantId}/tests`, { recruiterId: this.authService.getUserId(), testId: _testId, force: confirm }, { observe: 'response' });
   }
 }
